perf(web): cut render-blocking CSS and hydration work on the landing page (#1174)
Measured on a production build of the `/en` landing page (headless Chrome,
4x CPU throttle, ~1.6 Mbps, 7 interleaved rounds against both builds):

  LCP                    3700 ms -> 3464 ms   (-6.4%)
  main-thread long tasks  561 ms ->  450 ms   (-19.8%)
  render-blocking CSS    481.7 KB -> 325.7 KB (88.4 -> 63.4 KB gzip)
  first-load JS          871.0 KB -> 819.7 KB (263.7 -> 248.9 KB gzip)

Four separate causes:

1. Tailwind was compiled and shipped twice. Both globals.css and globals2.css
   started with `@import "tailwindcss"`, and the layout imported both, so the
   page loaded two near-identical 200 KB stylesheets. globals2.css is now
   imported from globals.css and the framework is emitted once.

2. Inter was declared twice via next/font (`--font-inter` and an unused
   `--font-heading`), producing two identical @font-face sets and duplicate
   font files. One instance now backs both; `--font-heading` was referenced
   nowhere in the stylesheets.

3. The whole 1340-line landing page was a single "use client" component, so
   every static section — hero, feature grids, programs, footer — was
   downloaded as JavaScript and hydrated before the page became interactive.
   The page is now a server component composing server-rendered sections plus
   small client islands for the parts that are genuinely interactive: header,
   screenshot gallery, contact form, newsletter form and the TestFlight
   dialog. Hero content, which holds the LCP element, no longer waits on the
   page bundle.

4. The features grid rendered six skeleton placeholders behind a hardcoded
   `setTimeout(..., 1500)` before swapping in nine static, already-known
   cards. The cards are now in the server-rendered HTML; this is the only
   change to the rendered output, verified by diffing the SSR markup of both
   builds (tag sequence, visible text and class lists are otherwise identical).

Also: the DNA cursor trail ran a requestAnimationFrame loop for the whole
session even while switched off, clearing a full-viewport canvas every frame.
The loop now only runs while the effect is enabled.

The unreachable "coming soon" modal is removed — its only trigger was a prop
threaded into AndroidStoreButtons, which never used it.
