+++ web/src/app/globals2.css
@@ -1,5 +1,5 @@
-@import "tailwindcss";
-
+/* Imported by globals.css — do not import "tailwindcss" here, it would compile
+   the framework a second time. */
 @custom-variant dark (&:is(.dark *));
 
 :root {
