+++ web/src/components/ui/navbar.tsx
@@ -2,7 +2,6 @@
 
 import Image from 'next/image'
 import Link from 'next/link'
-import '../../app/globals.css'
 
 import { PageWrapper } from '../layout'
